Rename IFrame to PaneWithWasmPack
